Delete unused files, Code cleaning #53 Note: try redis localhost
