+++ server/src/index.ts
@@ -7,10 +7,6 @@ import session from 'express-session'
 import proxy from 'express-http-proxy'
 import passport from 'passport'
 import { Strategy as KeycloakStrategy } from 'passport-keycloak-oauth2-oidc'
-import { ApolloServer } from 'apollo-server-express'
-
-import { mergeSchemaList } from '@utils'
-import { AppModule } from './modules'
 
 import { routes as addAuthRoutes } from './routes/auth'
 import { routes as addWebhooksRoutes } from './routes/webhooks'
@@ -20,7 +16,6 @@ import { setup as queueSetup } from './queue'
 import { logger, sessionStore } from '@shared'
 
 // API keys and Passport configuration
-// import * as passportConfig from './config/passport'
 
 const app = express()
 
@@ -48,25 +43,6 @@ app.use(sessionMiddleware)
 // }))
 app.use(passport.initialize())
 app.use(passport.session())
-
-// app.use((req, res, next) => {
-//   res.locals.user = req.user
-//   next()
-// })
-// app.use((req, res, next) => {
-//   // After successful login, redirect back to the intended page
-//   if (!req.user &&
-//     req.path !== '/login' &&
-//     req.path !== '/signup' &&
-//     !req.path.match(/^\/auth/) &&
-//     !req.path.match(/\./)) {
-//     req.session.returnTo = req.path
-//   } else if (req.user &&
-//     req.path == '/account') {
-//     req.session.returnTo = req.path
-//   }
-//   next()
-// })
 
 passport.serializeUser((user, done) => {
   done(null, user)
@@ -99,44 +75,11 @@ passport.use(
 
 async function main () {
   try {
-
-    // app.use('/v1/graphql', function (req, res, next) {
-    //   console.log('Request Type:', req.originalUrl, req.method, req.body)
-    //   next()
-    // })
-
-    const schema = await mergeSchemaList([
-      // 'http://localhost:8002/v1/graphql',
-      {
-        typeDefs: AppModule.typeDefs,
-        resolvers: AppModule.resolvers
-      }
-    ])
-
-    const server = new ApolloServer({
-      schema,
-      context: ({ req, res }) => {
-        return {
-          userId: 1
-        }
-      },
-      playground: {
-        settings: {
-          'editor.theme': 'light'
-        }
-      }
-    })
-
-    server.applyMiddleware({ app, path: '/v1/graphql' })
-
     addAuthRoutes(app, passport, sessionMiddleware, JSON.parse(process.env.USE_KEYCLOAK))
     addWebhooksRoutes(app, sessionStore, sessionMiddleware)
     addUploadRoutes(app, sessionMiddleware)
 
     app.use('/', proxy(process.env.APP_URL_PROXY))
-
-    // Auth routes
-    // app.use('/login', sessionMiddleware, authController.getlogin)
 
     // Auth callback
     await queueSetup()
