--- server/src/modules/index.ts
@@ -1,11 +1,0 @@
-import { GraphQLModule } from '@graphql-modules/core' // https://github.com/Urigo/graphql-modules
-import { AssetModule } from './asset'
-import { ProjectModule } from './project'
-
-export const AppModule = new GraphQLModule({
-  name: 'app',
-  imports: [
-    // AssetModule,
-    ProjectModule
-  ]
-})
