+++ server/src/shared/index.ts
@@ -1,3 +1,2 @@
 export * from './Logger'
-export * from '../config/redis'
-// export * from './minioClient'
+export * from './redis'
