+++ server/src/shared/redis.ts
@@ -1,11 +1,10 @@
 import session from 'express-session'
 import redis from 'redis'
 import redisStore from 'connect-redis'
-import { logger } from '../shared/Logger'
 
 let RedisStore = redisStore(session)
 let redisClient = redis.createClient()
-let store = new RedisStore({ host: '172.22.0.3', port: 6379, client: redisClient, ttl: 86400 })
+let store = new RedisStore({ host: 'localhost', port: 6379, client: redisClient, ttl: 86400 })
 
 export function getSessionUserId (sessionId: string) {
   return new Promise((resolve, reject) => {
