+++ server/src/utils/filesUtils.ts
@@ -1,0 +1,7 @@
+export interface IFileInfo {
+  size: number
+  md5: string
+  sha1: string
+  sha256: string
+  location?: string
+}
